pyproject.toml
[project]
name = "vision-winter-camp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.12.0.88",
    "ultralytics>=8.4.14",
    "torch==2.5.1+cu124",
    "torchvision==0.20.1+cu124",
    "qrcode[pil]>=8.2",
    "pyserial>=3.5",
    "pyzbar>=0.1.9",
]

# 这一块是关键：告诉 uv 去哪里找显卡版 torch
[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

# 这一块告诉 uv：torch 和 torchvision 必须去上面那个 pytorch-gpu 找
[tool.uv.sources]
torch = { index = "pytorch-gpu" }
torchvision = { index = "pytorch-gpu" }
